Classify Images

Classify Images

classify_img(
  classifier,
  path = NULL,
  img = NULL,
  feature_frame = NULL,
  filter_widths,
  shape_sizes,
  class_highlight = NULL,
  dims = NULL,
  preprocess_fun_img = NULL
)

Arguments

classifier
an object returned from ranger to use as a classifier

path
the path of the image to classify

img
an EBImage image

feature_frame
a data frame with the calculated features

filter_widths
filters widths to use when calculating features

class_highlight
classifier to show as white in the final mask

dims
dimensions of the images to classify, only required if using the feature_frmae interface

This classifier by default will fall back to file, trying in order of priority a `feature_frame` > `img` > `path`, please note that since the `feature_frame` does not contain the dimensions of the image, it has to be provided in the `dims` argument in addition.

Value

Examples

params_df <- tibble::tibble(
    file = c(
        system.file(
            "extdata", "tiny_4T1-shNT-1_layer1.png",
            package = "clasifierrr"),
        system.file(
            "extdata", "tiny_4T1-shNT-1_layer2.png",
            package = "clasifierrr")),
    classif = c("spheroid", "bg"),
    related_file = system.file(
        "extdata", "tiny_4T1-shNT-1.png",
        package = "clasifierrr")
    )
trainset <- build_train_multi(
    params_df, filter_widths = c(3, 5, 15),
    shape_sizes = c(21, 51))

#> Returning for file:  /tmp/Rtmpj1ueSh/temp_libpath734f364a36a4/clasifierrr/extdata/tiny_4T1-shNT-1_layer1.png and classification" spheroid " a total of { 8556 } positive pixels

#> Returning for file:  /tmp/Rtmpj1ueSh/temp_libpath734f364a36a4/clasifierrr/extdata/tiny_4T1-shNT-1_layer2.png and classification" bg " a total of { 14056 } positive pixels

#> Warning: The selected train size(50000) is larger than the number of classified pixels (22567)  so the number is getting updated to the total number of available pixels

#> Classified objects are of classes {bg: 14056} and {spheroid: 8511}
#> 

#> Returning a data frame of 22567 rows and 15 columns

trainset$pixel_class <- trainset$pixel_class == "spheroid"
model_simple_glm <- glm(pixel_class~.,data = trainset, family = binomial(link = "logit"))

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

class_img <- classify_img(
    model_simple_glm, path = params_df[[3]][[1]],
    filter_widths = c(3, 5, 15), shape_sizes = c(21, 51))

#> Attempting to read image from file: /tmp/Rtmpj1ueSh/temp_libpath734f364a36a4/clasifierrr/extdata/tiny_4T1-shNT-1.png

#> Attempting to calculate features

#> Starting classification

#> Warning: prediction from a rank-deficient fit may be misleading

#> Took  0.00914   secs  to predict the image

# plot(as.raster(class_img))