Calculates the difference of gaussian filter

This filter attempts to enhance the features in an image, it is based on subtracting a blurry image from a less blurry one. The ratio of this blurrines is measured by the ratio in their `sigma`.

dog_filter(img, width, ratio = 5, img_fft = fftwtools::fftw2d(img))

compile_dog_filter(width, dim_img, ratio = 5)

Arguments

img
Dimensions of the image

width
width of the filter to be applied

ratio
the ratio of the sigma between the blurrier and the less blurry images

img_fft
optional argument to pass the pre-computed fourier tranform of the image from `fftwtools::fftw2d(img)`.

dim_img
the dimensions of the image to be used later during filtering, only needed if pre-compiling the filter.

Value

Functions

compile_dog_filter: pre-calculates parameters and returns a function

See also

Examples

test_image <- matrix(runif(81), 9)
dog_filter(test_image, 3)

#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,] -1.255405e-08  3.255060e-08 -6.537995e-09  8.610646e-09  2.488961e-08
#>  [2,] -1.111771e-08  1.047991e-08 -1.372764e-08 -1.466059e-08 -2.312018e-08
#>  [3,]  2.483356e-08 -1.427798e-08 -1.618182e-08  2.818342e-08  7.282775e-09
#>  [4,]  6.753585e-09 -2.949175e-09 -1.542272e-09  1.735822e-08 -3.410949e-08
#>  [5,] -6.568204e-09 -1.558137e-08 -4.337539e-09  2.775658e-08 -2.082710e-09
#>  [6,]  1.412242e-08 -1.195444e-08  1.895442e-08 -2.904223e-08  1.738634e-08
#>  [7,] -1.647728e-08  2.440581e-08 -3.430613e-08  3.174243e-08  2.615773e-09
#>  [8,]  3.454089e-08 -2.400825e-08  2.011581e-08 -2.334523e-08 -2.759882e-08
#>  [9,] -1.722915e-08 -2.820226e-08  2.355118e-08  7.528716e-09  8.342799e-09
#>                [,6]          [,7]          [,8]          [,9]
#>  [1,]  1.201618e-08  2.784325e-09 -1.328384e-09  4.252790e-09
#>  [2,] -1.903805e-08  1.844528e-08 -6.150729e-09  1.936007e-08
#>  [3,] -6.638495e-09  1.728401e-09  1.168801e-08 -3.525838e-08
#>  [4,]  1.969540e-08 -2.343455e-09  2.634178e-09  7.687096e-09
#>  [5,] -2.350309e-08 -7.204522e-09  1.185584e-08  2.431728e-08
#>  [6,] -1.027446e-08  2.501608e-08 -2.780861e-08 -2.341593e-08
#>  [7,]  8.200783e-09 -9.091126e-09  2.401539e-08 -8.965694e-09
#>  [8,]  2.482031e-08 -2.433725e-09  1.684459e-08 -3.807407e-09
#>  [9,] -3.692205e-08  1.168425e-08 -8.216408e-09 -1.513873e-08

my_dog_filter <- compile_dog_filter(3, dim(test_image))
my_dog_filter(test_image)

#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,] -1.255405e-08  3.255060e-08 -6.537995e-09  8.610646e-09  2.488961e-08
#>  [2,] -1.111771e-08  1.047991e-08 -1.372764e-08 -1.466059e-08 -2.312018e-08
#>  [3,]  2.483356e-08 -1.427798e-08 -1.618182e-08  2.818342e-08  7.282775e-09
#>  [4,]  6.753585e-09 -2.949175e-09 -1.542272e-09  1.735822e-08 -3.410949e-08
#>  [5,] -6.568204e-09 -1.558137e-08 -4.337539e-09  2.775658e-08 -2.082710e-09
#>  [6,]  1.412242e-08 -1.195444e-08  1.895442e-08 -2.904223e-08  1.738634e-08
#>  [7,] -1.647728e-08  2.440581e-08 -3.430613e-08  3.174243e-08  2.615773e-09
#>  [8,]  3.454089e-08 -2.400825e-08  2.011581e-08 -2.334523e-08 -2.759882e-08
#>  [9,] -1.722915e-08 -2.820226e-08  2.355118e-08  7.528716e-09  8.342799e-09
#>                [,6]          [,7]          [,8]          [,9]
#>  [1,]  1.201618e-08  2.784325e-09 -1.328384e-09  4.252790e-09
#>  [2,] -1.903805e-08  1.844528e-08 -6.150729e-09  1.936007e-08
#>  [3,] -6.638495e-09  1.728401e-09  1.168801e-08 -3.525838e-08
#>  [4,]  1.969540e-08 -2.343455e-09  2.634178e-09  7.687096e-09
#>  [5,] -2.350309e-08 -7.204522e-09  1.185584e-08  2.431728e-08
#>  [6,] -1.027446e-08  2.501608e-08 -2.780861e-08 -2.341593e-08
#>  [7,]  8.200783e-09 -9.091126e-09  2.401539e-08 -8.965694e-09
#>  [8,]  2.482031e-08 -2.433725e-09  1.684459e-08 -3.807407e-09
#>  [9,] -3.692205e-08  1.168425e-08 -8.216408e-09 -1.513873e-08