Calculating sobel filters

The sobel filter is used to detect edges in the images

compile_sobel_filter(width, dim_img)

sobel_filter(img, width = 3, img_fft = fftwtools::fftw2d(img))

Arguments

width
width of the filter to be applied

dim_img
the dimensions of the image to be used later during filtering, only needed if pre-compiling the filter.

img
Dimensions of the image

img_fft
optional argument to pass the pre-computed fourier tranform of the image from `fftwtools::fftw2d(img)`.

Value

Functions

compile_sobel_filter: pre-calculates parameters and returns a function

See also

Examples

test_image <- matrix(runif(81), 9)
sobel_filter(test_image, 3)

#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4245455 2.2740550 0.9470254 1.6751756 0.5554362 1.4059110 1.5265315
#>  [2,] 0.5324474 1.8085346 0.9115098 1.8119673 0.2480912 1.2526553 2.0636632
#>  [3,] 0.3019312 1.6022725 0.9256606 1.1091678 1.0476057 0.1368977 1.4959338
#>  [4,] 0.7728303 2.2058616 1.2184364 1.0331480 1.0456269 1.4271508 1.2108049
#>  [5,] 0.9030726 2.0763441 1.2813699 0.6563363 0.4751827 1.5730021 1.7689434
#>  [6,] 1.5998222 1.2675711 1.0559480 1.2657187 1.5153197 0.1328465 0.9925235
#>  [7,] 1.1282111 2.2523007 2.2462812 1.1708179 0.4870297 1.3134240 0.8979116
#>  [8,] 0.3890807 0.8330062 1.1278124 0.9681000 2.0288845 1.8262824 2.5311444
#>  [9,] 0.4906121 1.2715136 2.2501113 0.7774343 2.1897731 2.3309071 1.9383280
#>            [,8]      [,9]
#>  [1,] 0.8809468 0.6626284
#>  [2,] 2.0014402 0.7694960
#>  [3,] 1.1077526 0.8251930
#>  [4,] 1.3126208 0.4717449
#>  [5,] 1.9308593 0.3848323
#>  [6,] 1.4091946 1.4347901
#>  [7,] 2.4225626 1.4140400
#>  [8,] 2.8395274 1.1664651
#>  [9,] 1.6446548 0.6834588

my_sobel_filter <- compile_sobel_filter(3, dim(test_image))
my_sobel_filter(test_image)

#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4245455 2.2740550 0.9470254 1.6751756 0.5554362 1.4059110 1.5265315
#>  [2,] 0.5324474 1.8085346 0.9115098 1.8119673 0.2480912 1.2526553 2.0636632
#>  [3,] 0.3019312 1.6022725 0.9256606 1.1091678 1.0476057 0.1368977 1.4959338
#>  [4,] 0.7728303 2.2058616 1.2184364 1.0331480 1.0456269 1.4271508 1.2108049
#>  [5,] 0.9030726 2.0763441 1.2813699 0.6563363 0.4751827 1.5730021 1.7689434
#>  [6,] 1.5998222 1.2675711 1.0559480 1.2657187 1.5153197 0.1328465 0.9925235
#>  [7,] 1.1282111 2.2523007 2.2462812 1.1708179 0.4870297 1.3134240 0.8979116
#>  [8,] 0.3890807 0.8330062 1.1278124 0.9681000 2.0288845 1.8262824 2.5311444
#>  [9,] 0.4906121 1.2715136 2.2501113 0.7774343 2.1897731 2.3309071 1.9383280
#>            [,8]      [,9]
#>  [1,] 0.8809468 0.6626284
#>  [2,] 2.0014402 0.7694960
#>  [3,] 1.1077526 0.8251930
#>  [4,] 1.3126208 0.4717449
#>  [5,] 1.9308593 0.3848323
#>  [6,] 1.4091946 1.4347901
#>  [7,] 2.4225626 1.4140400
#>  [8,] 2.8395274 1.1664651
#>  [9,] 1.6446548 0.6834588